Admin Institution - add protection against unauthorised ID change

+++ src/main/java/pl/coderslab/charity/services/impl/InstitutionServiceImpl.java
@@ -5,15 +5,13 @@ import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pl.coderslab.charity.domain.entities.Donation;
 import pl.coderslab.charity.domain.entities.Institution;
 import pl.coderslab.charity.domain.repositories.DonationRepository;
 import pl.coderslab.charity.domain.repositories.InstitutionRepository;
-import pl.coderslab.charity.dtos.DonationDataDTO;
 import pl.coderslab.charity.dtos.InstitutionAddDataDTO;
 import pl.coderslab.charity.dtos.InstitutionDataDTO;
+import pl.coderslab.charity.exceptions.EntityToDataBaseException;
 import pl.coderslab.charity.services.InstitutionService;
-import pl.coderslab.charity.services.SavingDataException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -69,7 +67,7 @@ public class InstitutionServiceImpl implements InstitutionService {
     }
 
     @Override
-    public void saveInstitution(InstitutionAddDataDTO institutionAddDataDTO) throws SavingDataException {
+    public void saveInstitution(InstitutionAddDataDTO institutionAddDataDTO) throws EntityToDataBaseException {
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! saveInstitution !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! ");
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.saveInstitution institutionDataDTO to be mapped to Institution: {}", institutionAddDataDTO.toString());
 
@@ -78,17 +76,22 @@ public class InstitutionServiceImpl implements InstitutionService {
         Institution institution = modelMapper.map(institutionAddDataDTO, Institution.class);
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.saveInstitution institution (from institutionDataDTO) after simple mapping: {}", institution.toString());
 
+        // Protection against unauthorised in fact update (instead of save=add new record/line)
+        if (institution.getId() != null) {
+            throw new EntityToDataBaseException("Wystąpił błąd przy walidacji lub zapisie danych. Powtórz całą operację");
+        }
+
         // Final saving donation
         Institution institutionSaved = institutionRepository.save(institution);
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.saveInstitution institutionSaved saved: {}", institutionSaved.toString());
         // check if saved
         if (institutionSaved == null) {
-            throw new SavingDataException("Wystąpił błąd przy zapisie danych. Powtórz całą operację");
+            throw new EntityToDataBaseException("Wystąpił błąd przy walidacji lub zapisie danych. Powtórz całą operację");
         }
     }
 
     @Override
-    public void updateInstitution(InstitutionDataDTO institutionDataDTO) throws SavingDataException {
+    public void updateInstitution(Long idProtected, InstitutionDataDTO institutionDataDTO) throws EntityToDataBaseException {
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! updateInstitution !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! ");
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.updateInstitution institutionDataDTO to be mapped to Institution: {}", institutionDataDTO.toString());
 
@@ -97,21 +100,26 @@ public class InstitutionServiceImpl implements InstitutionService {
         Institution institution = modelMapper.map(institutionDataDTO, Institution.class);
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.updateInstitution institution (from institutionDataDTO) after simple mapping: {}", institution.toString());
 
+        // Protection against unauthorised in fact update another record/line (instead of update the right one record/line)
+        if (institution.getId() != idProtected) {
+            throw new EntityToDataBaseException("Wystąpił błąd przy walidacji lub zapisie danych. Powtórz całą operację");
+        }
+
         // Final update donation
         Institution institutionSaved = institutionRepository.save(institution);
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.updateInstitution institutionSaved saved: {}", institutionSaved.toString());
         // check if update/saved
         if (institutionSaved == null) {
-            throw new SavingDataException("Wystąpił błąd przy zapisie danych. Powtórz całą operację");
+            throw new EntityToDataBaseException("Wystąpił błąd przy walidacji lub zapisie danych. Powtórz całą operację");
         }
         // check if updated (at update ID does not change)
         if (institution.getId() != null && institution.getId() != institutionSaved.getId()) {
-            throw new SavingDataException("Wystąpił błąd przy zapisie danych. Powtórz całą operację");
+            throw new EntityToDataBaseException("Wystąpił błąd przy walidacji lub zapisie danych. Powtórz całą operację");
         }
     }
 
     @Override
-    public void deleteInstitution(InstitutionDataDTO institutionDataDTO) {
+    public void deleteInstitution(Long idProtected, InstitutionDataDTO institutionDataDTO) throws EntityToDataBaseException {
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.deleteInstitution START !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! ");
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.deleteInstitution institutionDataDTO to be mapped to Institution: {}", institutionDataDTO.toString());
 
@@ -120,6 +128,10 @@ public class InstitutionServiceImpl implements InstitutionService {
         Institution institution = modelMapper.map(institutionDataDTO, Institution.class);
         log.debug("!!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! !!!!!!!!!!! InstitutionServiceImpl.deleteInstitution institution (from institutionDataDTO) after simple mapping: {}", institution.toString());
 
+        // Protection against unauthorised in fact delete another record/line (instead of delete the right one record/line)
+        if (institution.getId() != idProtected) {
+            throw new EntityToDataBaseException("Wystąpił błąd przy walidacji lub zapisie danych. Powtórz całą operację");
+        }
 
         // Deleting all related donations
         donationRepository.deleteAll(donationRepository.findAllWithCategoriesByInstitutionOrderByInstitution(institution));
